feat: implement engine core with SQL preprocessing and backend execution support
