+++ packages/engine/src/lib.rs
@@ -1,9 +1,9 @@
 mod backend;
+mod engine;
 mod error;
-mod lix;
 mod types;
 
 pub use backend::LixBackend;
+pub use engine::{boot, Engine};
 pub use error::LixError;
-pub use lix::{open_lix, Lix, OpenLixConfig};
-pub use types::{QueryResult, Value};
+pub use types::{Plan, QueryResult, Value};
