+++ plan.md
@@ -162,51 +162,38 @@ The `version` view reads from state cache with special handling for inheritance:
 
 ## Milestone 0: Rust engine core + SQLite/Postgres backends
 
-Goal: a Rust-native Lix core that can open either SQLite or Postgres and execute SQL
-through the engine pipeline (passthrough first, rewriting later).
+Goal: a Rust-native Lix core that preprocesses SQL and can drive either SQLite or
+Postgres via a backend. The engine stays sync at the boundary; async happens in
+SDKs and backends.
 
-### API Surface (Rust)
-
-```rust
-let lix = open_lix(OpenLixConfig {
-    backend: Box::new(SqliteBackend::new(SqliteConfig {
-        filename: ":memory:".into(),
-    })),
-})?;
-
-let _ = lix.execute("SELECT 1 + 1", &[])?;
-```
+### API Surface (Rust, engine)
 
 ```rust
-let lix = open_lix(OpenLixConfig {
-    backend: Box::new(PostgresBackend::new(PostgresConfig {
-        connection_string: "postgres://user:pass@localhost:5432/lix".into(),
-    })),
-})?;
-
-let _ = lix.execute("SELECT 1 + 1", &[])?;
+let engine = boot();
+let plan = engine.preprocess("SELECT 1 + 1", &[])?;
 ```
 
 ### Engine Contract (Rust)
 
-- `openLix(config)` returns a `Lix` handle bound to a backend.
-- `lix.execute(sql, params)`:
+- `boot()` returns an engine handle with sync methods.
+- `engine.preprocess(sql, params)`:
   - Parses SQL with `sqlparser-rs` (SQLite dialect).
-  - For Milestone 0: passthrough (no rewriting) and forwards to backend.
-  - Returns rows (for SELECT) and metadata (for mutations).
+  - For Milestone 0: passthrough (no rewriting).
+  - Returns a `Plan` (rewritten SQL + metadata for postprocess).
+- `engine.postprocess(plan, result)`:
+  - Updates engine state, hooks, caches as needed.
+  - For Milestone 0: no-op.
 
 ### Backend Strategy (mirrors JS Environment API)
 
-The JS SDK already routes all SQL through a `LixEnvironment` abstraction
-(`environment.call(...)`), keeping execution pluggable. Keep Rust aligned by
-requiring a backend object that implements a minimal contract, rather than an
-enum `kind` switch.
+The backend is host-provided and async-capable. SDKs orchestrate:
+`preprocess → backend.execute → postprocess`.
 
-Define a minimal backend trait that the Rust engine depends on:
+Define a minimal backend trait that SDKs depend on:
 
 ```rust
 pub trait LixBackend {
-    fn execute(&self, sql: &str, params: &[Value]) -> Result<QueryResult>;
+    async fn execute(&self, sql: &str, params: &[Value]) -> Result<QueryResult>;
 }
 ```
 
@@ -220,12 +207,34 @@ Implementations:
 
 ### Success Criteria
 
-1. `openLix()` works for SQLite and Postgres backends.
-2. `lix.execute()` runs `SELECT 1 + 1` on both backends.
+1. `boot()` can preprocess `SELECT 1 + 1` (passthrough).
+2. SDK `execute()` runs `SELECT 1 + 1` on both backends via preprocess→execute→postprocess.
 3. Parameter binding works for both backends (positional params).
 4. Transactions can be used via SQL (`BEGIN`, `COMMIT`, `ROLLBACK`) if needed.
 
-## Milestone 0.05: DataFusion SQL parse passthrough
+## Milestone 0.1: JS bindings (Node + WASM)
+
+Expose `openLix()` and `lix.execute()` to JS:
+
+Bindings consume the **WASM build of `packages/engine`** and keep the engine
+sync at the WASM boundary.
+
+- **Node**: WASM host + native JS backend (e.g. BetterSQLite3, node-postgres).
+- **Browser**: WASM host + browser SQLite backend (WASM SQLite or OPFS).
+- SDK `execute()` does `preprocess → backend.execute → postprocess`.
+
+## Milestone 0.2: Python bindings
+
+Expose `openLix()` and `lix.execute()` via WASM host:
+
+Bindings consume the **WASM build of `packages/engine`** and keep the engine
+sync at the WASM boundary.
+
+- Python wrapper matches the same API shape (backend objects).
+- Host uses native Python DB drivers (sqlite3 / psycopg) to implement `execute`.
+- SDK `execute()` does `preprocess → backend.execute → postprocess`.
+
+## Milestone 0.3: DataFusion SQL parse passthrough
 
 Hook in DataFusion’s SQL parser to parse incoming SQL, serialize it back to SQL,
 and forward the normalized SQL string to the backend. This keeps the execution
@@ -238,27 +247,6 @@ pipeline for later rewrites.
 2. Serialize AST back to SQL string (no rewrite yet)
 3. Forward serialized SQL to `LixBackend.execute`
 4. Add parity tests against raw passthrough for SQLite/Postgres
-
-## Milestone 0.1: JS bindings (Node + WASM)
-
-Expose `openLix()` and `lix.execute()` to JS:
-
-Bindings consume the **WASM build of `packages/engine`**.
-
-- **Node**: WASM host + native JS backend (e.g. BetterSQLite3, node-postgres).
-- **Browser**: WASM host + browser SQLite backend (WASM SQLite or OPFS).
-- Same API shape as Milestone 0, with `backend` as a class instance.
-  - `new SqliteBackend(...)` first.
-  - `new PostgresBackend(...)` optional.
-
-## Milestone 0.2: Python bindings
-
-Expose `openLix()` and `lix.execute()` via PyO3:
-
-Bindings consume the **WASM build of `packages/engine`**.
-
-- Python wrapper matches the same API shape (backend objects).
-- Host uses native Python DB drivers (sqlite3 / psycopg) to implement `execute`.
 
 ## Milestone 1: Untracked State
